Stop using _flag variables directly
By creating a variable e.g. just without the _flag suffix;
one can change the actions depending on the other flags in the
new function:
    _post_handle_flags()

This can make the rest of the code easier to read by avoiding
complex if statements with multiple different flags and instead
only have one variable which it checks. Can be a variable which
by it self doesn't have a flag, but which can be created by e.g. a
combination of flags.